requirements.txt
aiohttp
disnake
PyYAML